requirements.txt
pydub==0.25.1
python-telegram-bot==13.5
requests==2.25.1
